Fix for jgit running on the Harmony JRE
Under Harmony, the Inflator.inflate() method will return immediately without examining the stream if you ask for zero-bytes of data.

This means finished() will never return true if you are naively trying to inflate a zero-length file into a zero-length array....

....jgit unpacking a zero-length file blob  will consequently blow up

"fatal: Packfile corruption detected: Unknown zlib error."

The 1 byte array used by the fix should never get populated, if it does, that's a proper cause to raise an Exception. So we care not for the contents of the array, or concurrent access to it. Let everyone share the same copy of this 1-byte array.