Enable full Transport configuration for JGit API commands
Add a TransportConfigCallback parameter to JGit API commands to allow
transport-type-specific configuration to be set on the Transport
after it's been created.

My specific use-case is configuring additional properties on
SshTransport - I need to take over the SshSessionFactory used by the
transport. Using the TransportConfigCallback parameter I can simply
do this (rather than reimplement the commands):

public void configure(Transport tn) {
  if (tn instanceof SshTransport) {
    ((SshTransport) tn).setSshSessionFactory(factoryProvider.get());
  }
}

The Transport type is only known once the remote has been resolved
and the URI parsed, which is why it makes sense to have this as a
callback- passing the generated Transport object back to the user
allows them to easily configure it without restriction.

Alternatives to this approach:

- Specifically add a SshSessionFactory parameter to JGit API
commands, which would add nasty transport-specific code to the API
and set a bad precedent for ballooning the API with other transport-
specific parameters.

- Allow users to pre-create the Transport and set it on the command.
Not good: the user has to resolve the remote for themselves.

- A non-static Transport-factory approach, letting user supply the
command a TransportFactory. A larger change (removing all the static
'open()' methods from Transport) and more than I need for this use
case at least - I'm happy enough with the transport resolution, I
just want to configure the transport once it's generated.

-so the callback is the best option in my view.

A note about where this leaves the API - there are now 4 commands:

CloneCommand
PullCommand
FetchCommand
PushCommand

-that share 3 identical transport-related parameters:

timeout
credentialsProvider
transportConfigurator

I think there's potential for introducing an interface or val-object
to identify/encapsulate this repetition, which I'd be happy to do in
a subsequent commit.

Change-Id: I8983c3627cdd7d7b2aeb0b6a3dadee553378b951