Fix for broken Android implementation of InflaterInputStream.close()
This fix is a bit of a hack - low-impact, but also fixing the problem in
the 'wrong' place, ie the Inflaters used by the InflaterCache- really,
should be fixing the implementation of InflaterInputStream. On Android
InflaterInputStream always ends() the Inflater it was using when closed,
even if the Inflater was not internally constructed ut supplied in the
constructor - in Oracle Java externally-provided Inflaters are NOT ended.

See https://github.com/rtyley/agit/issues/47 for more details.